refactor(theme): adjust theme config & fonts
