+++ package.json
@@ -18,9 +18,6 @@
     "@chakra-ui/react": "2.4.2",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
-    "@fontsource/abril-fatface": "^4.5.7",
-    "@fontsource/caveat": "^4.5.7",
-    "@fontsource/inter": "^4.5.11",
     "@types/crypto-js": "^4.1.1",
     "axios": "^1.2.1",
     "crypto-js": "^4.1.1",
