+++ src/pages/index.tsx
@@ -9,13 +9,13 @@ import Link from "next/link";
 
 const Home = () => {
   return (
-    <Grid gap={12} marginY={[8, 16]} w="full" textAlign="center">
-      <Heading letterSpacing={1} fontWeight="extrabold">
+    <Grid gap={8} marginY={[8, 16]} w="full" textAlign="center">
+      <Heading letterSpacing={1} size="2xl" fontWeight="extrabold">
         Create Greetings for your Folks!
       </Heading>
       <Image
         marginX="auto"
-        width={[240, 320, 400]}
+        width={[200, 240, 280]}
         src="/High five-cuate.svg"
         alt="illustration"
       />
@@ -23,7 +23,13 @@ const Home = () => {
         Illustration by Freepik Storyset
       </ChakraLink>
 
-      <Button as={Link} href="/create" size="lg" marginX={[0, 16, 32]}>
+      <Button
+        as={Link}
+        href="/create"
+        size="lg"
+        marginX={[0, 16, 32]}
+        colorScheme="blue"
+      >
         Create One
       </Button>
     </Grid>
