--- src/styles/customTheme.ts
@@ -1,31 +1,0 @@
-import { theme, extendTheme } from "@chakra-ui/react";
-
-const customTheme = extendTheme({
-  fonts: {
-    ...theme.fonts,
-    heading: "Abril Fatface, serif",
-    body: "Inter, sans-serif",
-  },
-  colors: {
-    ...theme.colors,
-    /** Example */
-    // teal: {
-    //   ...theme.colors.teal,
-    //   700: "#005661",
-    //   500: "#00838e",
-    //   300: "#4fb3be",
-    // },
-  },
-  components: {
-    Button: {
-      baseStyle: {
-        borderRadius: 24,
-      },
-    },
-  },
-  config: {
-    disableTransitionOnChange: false,
-  },
-});
-
-export default customTheme;
