+++ src/styles/theme/components/button.ts
@@ -1,0 +1,7 @@
+import type { ComponentStyleConfig } from "@chakra-ui/react";
+
+export const Button: ComponentStyleConfig = {
+  baseStyle: {
+    borderRadius: 24,
+  },
+};
