+++ src/styles/theme/components/index.ts
@@ -1,0 +1,5 @@
+import { Button } from "./button";
+
+export const components = {
+  Button,
+};
